iers2py: Update how NIST and IERS data are merged, again
The previous algorithm looked for a date where NIST and IERS matched,
and then switched to IERS.  However, we're in a period of time where
NIST has continued broadcasting -200 even while IERS data went as low as
about -255 (before going back to about -180), which this script rounded
to -300.

Instead, use the timestamp of modification of the NIST HTML document as
the date to switch to IERS data.