Add wwvbdec
wwvbdec.wwvbreceive is a stateful decoder of the amplitude portion of the
wwvb signal implemented as a generator coroutine.

Given a sequence consisting of the MARK of second 59 of a previous
minute, plus the 60 seconds of the current minute, it will yield
a WWVBTimecode.  At other times, it will yield None.

The returned timecode will contain 60 items, even if the minute contains
a leap second and there is nominally an extra MARK within the minute.

Low level validation (must be mark / must be zero / DUT1 sign bits)
is performed, but high level validation (e.g., day number is in range)
is not.  I defend this decision because otherwise e.g., you must poke
into this routine knowledge of what the WWVB epoch is so that you know
whether year 00 has a day 366 or not.