welcome to doc dependency hell
sphinx-rtd-theme 2.0.0 declares that it is incompatible with sphinx 8.
As a result, with the old requirements-dev.txt, sphinx 8 and
sphinx-rtd-theme 0.mumble are installed, as sphinx-rtd-theme 0.mumble did
NOT declare an explicit incompatibility with sphinx 8.

This pairing emphatically does NOT work, giving the error
`UndefinedError("'style' is undefined")` during doc building.

For all these AWESOME packages, specify a dependency range.
This is sure to prevent future breakage :wink: